#705: add CacheLoaderWriterConfiguration interface + move write behind classes under o.e.c.loaderwriter.writebehind
