+++ core/src/main/java/org/ehcache/config/loaderwriter/CacheLoaderWriterConfiguration.java
@@ -1,0 +1,10 @@
+package org.ehcache.config.loaderwriter;
+
+import org.ehcache.spi.loaderwriter.CacheLoaderWriterProvider;
+import org.ehcache.spi.service.ServiceConfiguration;
+
+/**
+ * @author Ludovic Orban
+ */
+public interface CacheLoaderWriterConfiguration extends ServiceConfiguration<CacheLoaderWriterProvider> {
+}
