+++ core/src/main/java/org/ehcache/config/loaderwriter/writebehind/DefaultWriteBehindConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ehcache.config.writebehind;
+package org.ehcache.config.loaderwriter.writebehind;
 
 import org.ehcache.spi.loaderwriter.WriteBehindConfiguration;
 import org.ehcache.spi.loaderwriter.WriteBehindProvider;
