+++ impl/src/main/java/org/ehcache/config/loaderwriter/DefaultCacheLoaderWriterConfiguration.java
@@ -19,12 +19,11 @@ package org.ehcache.config.loaderwriter;
 import org.ehcache.internal.classes.ClassInstanceConfiguration;
 import org.ehcache.spi.loaderwriter.CacheLoaderWriter;
 import org.ehcache.spi.loaderwriter.CacheLoaderWriterProvider;
-import org.ehcache.spi.service.ServiceConfiguration;
 
 /**
 * @author Alex Snaps
 */
-public class DefaultCacheLoaderWriterConfiguration extends ClassInstanceConfiguration<CacheLoaderWriter<?, ?>> implements ServiceConfiguration<CacheLoaderWriterProvider> {
+public class DefaultCacheLoaderWriterConfiguration extends ClassInstanceConfiguration<CacheLoaderWriter<?, ?>> implements CacheLoaderWriterConfiguration {
 
   public DefaultCacheLoaderWriterConfiguration(final Class<? extends CacheLoaderWriter<?, ?>> clazz, Object... arguments) {
     super(clazz, arguments);
